+++ src/_common/theme/theme-editor/theme-editor.vue
@@ -5,7 +5,7 @@ import { Api } from '../../api/api.service';
 import AppColorpicker from '../../colorpicker/AppColorpicker.vue';
 import AppLoading from '../../loading/AppLoading.vue';
 import { SiteTemplateModel } from '../../site/template/template-model';
-import AppThemeEditorFontSelector from './font-selector.vue';
+import AppThemeEditorFontSelector from './AppThemeEditorFontSelector.vue';
 import AppThemeEditorImage from './image.vue';
 
 interface StyleGroup {
